Tie MLeap version to Spark version.
