+++ mleap-spark/src/main/resources/reference.conf
@@ -1,51 +1,4 @@
-ml.combust.mleap.spark.registry.v20-ops = [
-  "org.apache.spark.ml.bundle.ops.classification.GBTClassifierOpV20",
-  "org.apache.spark.ml.bundle.ops.classification.LogisticRegressionOpV20",
-  "org.apache.spark.ml.bundle.ops.feature.StringIndexerOpV22",
-  "org.apache.spark.ml.bundle.ops.feature.OneHotEncoderOp"
-]
-
-ml.combust.mleap.spark.registry.v21-ops = [
-  "org.apache.spark.ml.bundle.ops.feature.MinHashLSHOp",
-  "org.apache.spark.ml.bundle.ops.feature.BucketedRandomProjectionLSHOp",
-  "org.apache.spark.ml.bundle.ops.classification.GBTClassifierOpV20",
-  "org.apache.spark.ml.bundle.ops.classification.LogisticRegressionOpV21",
-  "org.apache.spark.ml.bundle.ops.feature.StringIndexerOpV22",
-  "org.apache.spark.ml.bundle.ops.feature.OneHotEncoderOp"
-]
-
-ml.combust.mleap.spark.registry.v22-ops = [
-  "org.apache.spark.ml.bundle.ops.feature.MinHashLSHOp",
-  "org.apache.spark.ml.bundle.ops.feature.BucketedRandomProjectionLSHOp",
-  "org.apache.spark.ml.bundle.ops.classification.GBTClassifierOpV22",
-  "org.apache.spark.ml.bundle.ops.classification.LogisticRegressionOpV21",
-  "org.apache.spark.ml.bundle.ops.feature.StringIndexerOpV22",
-  "org.apache.spark.ml.bundle.ops.feature.OneHotEncoderOp"
-]
-
-ml.combust.mleap.spark.registry.v23-ops = [
-  "org.apache.spark.ml.bundle.ops.feature.MinHashLSHOp",
-  "org.apache.spark.ml.bundle.ops.feature.BucketedRandomProjectionLSHOp",
-  "org.apache.spark.ml.bundle.ops.classification.GBTClassifierOpV22",
-  "org.apache.spark.ml.bundle.ops.classification.LogisticRegressionOpV21",
-  "org.apache.spark.ml.bundle.ops.feature.FeatureHasherOp",
-  "org.apache.spark.ml.bundle.ops.feature.StringIndexerOpV23",
-  "org.apache.spark.ml.bundle.ops.feature.OneHotEncoderOpV23"
-]
-
-ml.combust.mleap.spark.registry.v20.ops += "ml.combust.mleap.spark.registry.base-ops"
-ml.combust.mleap.spark.registry.v20.ops += "ml.combust.mleap.spark.registry.v20-ops"
-
-ml.combust.mleap.spark.registry.v21.ops += "ml.combust.mleap.spark.registry.base-ops"
-ml.combust.mleap.spark.registry.v21.ops += "ml.combust.mleap.spark.registry.v21-ops"
-
-ml.combust.mleap.spark.registry.v22.ops += "ml.combust.mleap.spark.registry.base-ops"
-ml.combust.mleap.spark.registry.v22.ops += "ml.combust.mleap.spark.registry.v22-ops"
-
-ml.combust.mleap.spark.registry.v23.ops += "ml.combust.mleap.spark.registry.base-ops"
-ml.combust.mleap.spark.registry.v23.ops += "ml.combust.mleap.spark.registry.v23-ops"
-
-ml.combust.mleap.spark.registry.base-ops = [
+ml.combust.mleap.spark.registry.builtin-ops = [
   "org.apache.spark.ml.bundle.ops.classification.DecisionTreeClassifierOp",
   "org.apache.spark.ml.bundle.ops.classification.NaiveBayesClassifierOp",
   "org.apache.spark.ml.bundle.ops.classification.MultiLayerPerceptronClassifierOp",
@@ -95,5 +48,15 @@ ml.combust.mleap.spark.registry.base-ops = [
   "org.apache.spark.ml.bundle.ops.tuning.CrossValidatorOp",
   "org.apache.spark.ml.bundle.ops.tuning.TrainValidationSplitOp",
 
+  "org.apache.spark.ml.bundle.ops.feature.MinHashLSHOp",
+  "org.apache.spark.ml.bundle.ops.feature.BucketedRandomProjectionLSHOp",
+  "org.apache.spark.ml.bundle.ops.classification.GBTClassifierOp",
+  "org.apache.spark.ml.bundle.ops.classification.LogisticRegressionOp",
+  "org.apache.spark.ml.bundle.ops.feature.FeatureHasherOp",
+  "org.apache.spark.ml.bundle.ops.feature.StringIndexerOp",
+  "org.apache.spark.ml.bundle.ops.feature.OneHotEncoderOp",
+
   "org.apache.spark.ml.bundle.ops.PipelineOp"
 ]
+
+ml.combust.mleap.spark.registry.default.ops += "ml.combust.mleap.spark.registry.builtin-ops"
